refactor: enhance linkable extension trait
